+++ et-driver/CHANGELOG.md
@@ -11,6 +11,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Changed
 ### Deprecated
 ### Removed
+- [SW-9608] Removed legacy DMA read/write support
 ### Fixed
 ### Security